# openapi.json
{
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
            "Response": {
                "properties": {
                    "response": {
                        "title": "Response",
                        "type": "string"
                    }
                },
                "required": [
                    "response"
                ],
                "title": "Response",
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            }
        }
    },
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "openapi": "3.0.2",
    "paths": {
        "/get_order_line/{header_id}": {
            "get": {
                "description": "Retrieve order lines for a specific order.",
                "operationId": "get_order_line_get_order_line__header_id__get",
                "parameters": [
                    {
                        "in": "path",
                        "name": "header_id",
                        "required": true,
                        "schema": {
                            "title": "Header Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Get Order Line"
            }
        },
        "/get_order_lines": {
            "get": {
                "description": "Retrieve order lines for a specific order.",
                "operationId": "get_order_lines_get_order_lines_get",
                "parameters": [
                    {
                        "in": "query",
                        "name": "header_id",
                        "required": true,
                        "schema": {
                            "title": "Header Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Get Order Lines"
            }
        },
        "/get_orders": {
            "get": {
                "description": "Retrieve all orders from the database.",
                "operationId": "get_orders_get_orders_get",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                }
                            }
                        },
                        "description": "Successful Response"
                    }
                },
                "summary": "Get Orders"
            }
        },
        "/process-prompt": {
            "post": {
                "operationId": "process_prompt_process_prompt_post",
                "parameters": [
                    {
                        "in": "query",
                        "name": "prompt",
                        "required": true,
                        "schema": {
                            "title": "Prompt",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "Params",
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Response"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Process Prompt"
            }
        }
    }
}